Data cleaning complete
Filter the Birth Years column so only reasonable ages remain because the original maximum age in the data was 145 years.
Save only the September data to a CSV file so a small extract of data can be used to build the Tableau dashboard before importing the aggregated data to avoid performance issues.
Update the README by adding NumPy to the Technologies section.