Start Tableau story
Aggregate the first 6 months of Citi Bike's 2019 data to get a better idea of what the Tableau visualizations look like before using the full 12 months to avoid performance issues.
Complete the Tableau visualizations and place them in the dashboards for the Tableau story.
Add the Tableau Public link to the README to view the project's progress.