Project complete
Add a seasonal filter to each dashboard to separate the data by summer or winter.
Complete the dashboards and use them to create a story.
Add a link to the completed project in Tableau Public to the README so viewers can explore the dashboards.
Change the 'Project Status' in the README to 'Completed.'