Project complete
Modify the code to automatically update the calculated ages by using the datetime module to calculate the current year every time the code is run. 'datetime.now().year' replaces the written '2019' so that the oldest and youngest ages are always calculated based on the current year rather than only 2019.
The 'datetime.now().year' code is included as a comment to make the current year calculation optional since the Tableau data story is based on the year 2019.
Add 'Datetime' to the Technologies section of the README since the datetime module had to be imported to update the code.